initial refactor of dynamodb progress tracking into a pluggable backend
