+++ tests/test_interface.py
@@ -1,5 +1,6 @@
 import pytest
-from watchbot_progress import create_job, Part, WatchbotProgress, JobFailed
+from watchbot_progress import create_job, Part, JobFailed
+from watchbot_progress.backends.base import WatchbotProgressBase
 from mock import patch
 
 parts = [
@@ -8,7 +9,7 @@ parts = [
     {'source': 'c.tif'}]
 
 
-class MockBase(WatchbotProgress):
+class MockProgress(WatchbotProgressBase):
 
     def status(self, x):
         return {'progress': 0.3}
@@ -33,78 +34,61 @@ def test_create_jobs(monkeypatch):
     monkeypatch.setenv('WorkTopic', 'abc123')
     monkeypatch.setenv('ProgressTable', 'arn::table/foo')
 
-    class MockProgress(MockBase):
-        pass
-
-    with patch('watchbot_progress.uuid.uuid4', new=lambda: '000-123'):
-        with patch('watchbot_progress.WatchbotProgress', new=MockProgress):
-            assert create_job(parts) == '000-123'
+    with patch('uuid.uuid4', new=lambda: '000-123'):
+        assert create_job(parts, progress=MockProgress()) == '000-123'
 
 
 def test_Part_job_not_done(monkeypatch):
     monkeypatch.setenv('WorkTopic', 'abc123')
     monkeypatch.setenv('ProgressTable', 'arn::table/foo')
 
-    class MockProgress(MockBase):
-        def status(self, x):
-            return {'progress': 0.3}
-
-    with patch('watchbot_progress.WatchbotProgress', new=MockProgress):
-        msg = {'source': 'a.tif', 'partid': 1, 'jobid': '123'}
-        with Part(**msg):
-            msg['source']
+    with Part(jobid='123', partid=1, progress=MockProgress()):
+        pass
 
 
 def test_Part_job_done(monkeypatch):
     monkeypatch.setenv('WorkTopic', 'abc123')
     monkeypatch.setenv('ProgressTable', 'arn::table/foo')
 
-    class MockProgress(MockBase):
+    class CustomProgress(MockProgress):
         def complete_part(self, jobid, partid):
             return True
 
-    with patch('watchbot_progress.WatchbotProgress', new=MockProgress):
-        msg = {'source': 'a.tif', 'partid': 1, 'jobid': '123'}
-        with Part(**msg):
-            msg['source']
+    with Part(jobid='123', partid=1, progress=CustomProgress()):
+        pass
 
 
 def test_Part_already_failed(monkeypatch):
     monkeypatch.setenv('WorkTopic', 'abc123')
     monkeypatch.setenv('ProgressTable', 'arn::table/foo')
 
-    class MockProgress(MockBase):
+    class CustomProgress(MockProgress):
         def status(self, x):
             return {'progress': 0.3, 'failed': True}
 
     class CustomException(Exception):
         pass
 
-    with patch('watchbot_progress.WatchbotProgress', new=MockProgress):
-        msg = {'source': 'a.tif', 'partid': 1, 'jobid': '123'}
-        with pytest.raises(JobFailed) as e:
-            with Part(fail_job_on=[CustomException], **msg):
-                # sees that the overall job failed and will not execute this code
-                raise NotImplementedError("You'll never get here")
-        assert 'already failed' in str(e)
+    msg = {'source': 'a.tif', 'partid': 1, 'jobid': '123'}
+    with pytest.raises(JobFailed) as e:
+        with Part(fail_job_on=[CustomException], progress=CustomProgress(), **msg):
+            # sees that the overall job failed and will not execute this code
+            raise NotImplementedError("You'll never get here")
+    assert 'already failed' in str(e)
 
 
 def test_create_jobs_metadata(monkeypatch):
     monkeypatch.setenv('WorkTopic', 'abc123')
     monkeypatch.setenv('ProgressTable', 'arn::table/foo')
 
-    class MockProgress(MockBase):
-        pass
-
     meta = {'foo': 'bar'}
 
-    with patch('watchbot_progress.uuid.uuid4', new=lambda: '000-123'):
-        with patch('watchbot_progress.WatchbotProgress', new=MockProgress):
-            assert create_job(parts, metadata=meta) == '000-123'
+    with patch('uuid.uuid4', new=lambda: '000-123'):
+        assert create_job(parts, progress=MockProgress(), metadata=meta) == '000-123'
 
 
-@patch('watchbot_progress.WatchbotProgress.fail_job')
-@patch('watchbot_progress.WatchbotProgress.status')
+@patch('watchbot_progress.main.DynamoProgress.fail_job')
+@patch('watchbot_progress.main.DynamoProgress.status')
 def test_Part_fail_job_on(status, fail_job, monkeypatch):
     monkeypatch.setenv('WorkTopic', 'abc123')
     monkeypatch.setenv('ProgressTable', 'arn::table/foo')
@@ -119,8 +103,8 @@ def test_Part_fail_job_on(status, fail_job, monkeypatch):
     fail_job.assert_called_once_with(2, 2)
 
 
-@patch('watchbot_progress.WatchbotProgress.fail_job')
-@patch('watchbot_progress.WatchbotProgress.status')
+@patch('watchbot_progress.main.DynamoProgress.fail_job')
+@patch('watchbot_progress.main.DynamoProgress.status')
 def test_Part_dont_fail_job_on(status, fail_job, monkeypatch):
     monkeypatch.setenv('WorkTopic', 'abc123')
     monkeypatch.setenv('ProgressTable', 'arn::table/foo')
